Fix KafkaProducer#flush() behavior
A producer batch is split into multiple smaller batches when the client recieves Erros.MESSAGE_TOO_LARGE code from the broker. The future's returned as part of KafkaProducer#send() api are chained to make sure that it isn't evaluated until the newly created small batches are sent to the server. This chaining behavior is lacking in KafkaProducer#flush() api which breaks the api contract that all the future's will be in done state when the flush is complete.